+++ core/comps/contacts/srvc-contact-model.js
@@ -8,19 +8,22 @@ angular.module('cmContacts')
     'cmStateManagement',
     'cmUtil',
     'cmLogger',
+    'cmSecurityAspectsContact',
     '$q',
-    function(cmContactsAdapter, cmIdentityFactory, cmObject, cmStateManagement, cmUtil, cmLogger, $q){
+    function(cmContactsAdapter, cmIdentityFactory, cmObject, cmStateManagement, cmUtil, cmLogger, cmSecurityAspectsContact, $q){
         function ContactModel(data){
             var self = this;
 
             cmObject.addEventHandlingTo(this);
 
-            this.state  = new cmStateManagement(['loading']);
+            this.state         = new cmStateManagement(['loading']);
 
             this.id            = undefined;
             this.contactType   = undefined;
             this.group         = [];
             this.identity      = cmIdentityFactory.new();
+
+            this.securityAspects    = new cmSecurityAspectsContact(this);
 
             function init(data){
                 //cmLogger.debug('cmContactModel:init');
@@ -108,6 +111,10 @@ angular.module('cmContacts')
             };
 
             init(data);
+
+            this.identity.on('update:finished', function(){
+                self.securityAspects.scheduleRefresh();
+            });
 
         }