+++ core/comps/conversations/fcty-security_aspects_conversation.js
@@ -1,6 +1,6 @@
 'use strict';
 
-angular.module('cmSecurityAspects')
+angular.module('cmConversations')
 .factory('cmSecurityAspectsConversation',[
 
     'cmSecurityAspects',
@@ -95,37 +95,35 @@ angular.module('cmSecurityAspects')
                     dependencies: ['ENCRYPTED', 'NO_SYMMETRIC_KEY_TRANSMISSION'],
                     value: 1,
                     check: function(conversation){
-                        /**
-                         * @todo work around hack from BB 19.09.2014
-                         * new workaround for promises 17.10.2014 AP
-                         */
-                        if(conversation.recipients.length < 3){
-                            $q.all(conversation.recipients.map(function(recipient){
-                                return cmUserModel.verifyTrust(recipient, true)
-                            }))
-                            .then(
-                                function(){
 
-                                    if(!conversation.workaround_aspects_trusted){
-                                        conversation.workaround_aspects_trusted = true
-                                        self.refresh()
-                                    }
+                        //temporary solution, AP
+                        return  conversation.recipients.length < 3
+                            ?   conversation.recipients.reduce(function(so_far, recipient){
+                                    return  so_far
+                                            .then(function(){
+                                                return  cmUserModel.verifyIdentityKeys(recipient, true)
+                                            })
+                                            .then(function(ttrusted_keys){
+                                                return  ttrusted_keys.length == recipient.keys.length
+                                                        ?   $q.when(true)
+                                                        :   $q.reject()
+                                            })
 
-                                },
-                                function(){
-                                    if(conversation.workaround_aspects_trusted){
-                                        conversation.workaround_aspects_trusted = false
-                                        self.refresh()
-                                    }
-
-                                }
-                            )
-                            return conversation.workaround_aspects_trusted
-                        } else {
-                            return false
-                        }
+                                }, $q.when())
+                                .then(
+                                    function(){ return true },
+                                    function(){ return false }
+                                )
+                            :   $q.when(false)
                     }
                 });
+
+            
+
+            conversation.on('update:finished encryption:enabled encryption:disabled captcha:enabled captcha:disabled aspects:added', self.scheduleRefresh);
+            conversation.recipients.on('register update:finished deregister', self.scheduleRefresh);
+            cmUserModel.on('key:stored key:removed cache:updated', self.scheduleRefresh);
+
 
             return self;
         }