+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 ## Dev
+# WW 12.12.2014
+* new config.xml handling for phonegap
+** local building leads www/res/.. at path to resources
+** build.phonegap building needs res/.. at path to resources
+** new config with plugins and resoures for phonegap config/cameoConfig-phonegap.json
+
+# WW 11.12.2014
+* trusted contacts
+** in contact list, detail and recipients
+** oo = check trust
+** oo! = no key and untrusted
+** xo! = has key but untrusted
+** xx = trusted connection between two contacts
+* unit test command line -specs option available
+* fix icons for anrdoid devices lower then kitkat < 4.4
+
 # BS 11.12.2014
 * fix CAM-1490, add email or phone number to possible display names